Move EffectBase into dedicated header file
