+++ UI/EncyclopediaDetailPanel.cpp
@@ -9,6 +9,7 @@
 #include "CUILinkTextBlock.h"
 #include "MapWnd.h"
 #include "../universe/ConditionBase.h"
+#include "../universe/EffectAccounting.h"
 #include "../universe/ValueRefBase.h"
 #include "../universe/Encyclopedia.h"
 #include "../universe/Universe.h"
@@ -21,7 +22,6 @@
 #include "../universe/Special.h"
 #include "../universe/Species.h"
 #include "../universe/Field.h"
-#include "../universe/Effect.h"
 #include "../universe/Predicates.h"
 #include "../Empire/Empire.h"
 #include "../Empire/EmpireManager.h"
