+++ parse/CommonParamsParser.cpp
@@ -10,6 +10,7 @@
 #include "ValueRefParser.h"
 
 #include "../universe/Condition.h"
+#include "../universe/EffectAccounting.h"
 #include "../universe/ValueRef.h"
 
 #include <boost/spirit/include/phoenix.hpp>
