+++ parse/EffectParser.cpp
@@ -7,7 +7,7 @@
 #include "EffectParser5.h"
 
 #include "../universe/ConditionBase.h"
-#include "../universe/Effect.h"
+#include "../universe/EffectAccounting.h"
 
 #include <boost/spirit/include/phoenix.hpp>
 //TODO: replace with std::make_unique when transitioning to C++14
