+++ universe/Building.cpp
@@ -1,7 +1,7 @@
 #include "Building.h"
 
-#include "Effect.h"
 #include "ConditionBase.h"
+#include "EffectAccounting.h"
 #include "Planet.h"
 #include "Predicates.h"
 #include "Universe.h"
