+++ universe/Universe.cpp
@@ -12,6 +12,7 @@
 #include "../Empire/EmpireManager.h"
 #include "IDAllocator.h"
 #include "Building.h"
+#include "EffectAccounting.h"
 #include "Fleet.h"
 #include "Planet.h"
 #include "Ship.h"
@@ -19,7 +20,6 @@
 #include "Field.h"
 #include "UniverseObject.h"
 #include "UniverseGenerator.h"
-#include "Effect.h"
 #include "Predicates.h"
 #include "Special.h"
 #include "Species.h"
